Control de variables en clases de mapeo con sets y comprobaciones, mejorar la lógica de inserción y actualización en formularios de administración separando la logica del boton y haciendo las comprobaciones necesarias.

+++ src/main/java/Vista/Admin/Tablas/GestionHistorialAcademicoAdmin.java
@@ -287,7 +287,7 @@ public class GestionHistorialAcademicoAdmin extends JPanel {
             Object[] fila = {
                     historial.getEstudiante().getNombre() + " " + historial.getEstudiante().getApellido(),
                     historial.getAsignatura().getNombre(),
-                    (Double) historial.getNotaFinal(),
+                    historial.getNotaFinal(),
                     historial.getFechaAprobacion().toString(),
                     historial.getComentarios(),
                     historial
